// src/app/models/request.model.ts
export interface Request{
    id_application?: number;
    id_company? : number;
    description?: string;
    date_application?: string;
    hour_application?: string;
    type_application? : string;
    status_application?: string;
    company_name?: string;
    main_email?: string;
    number_phone?: string;
    slug?: string;
}